+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.21
 ------------
 - improvements to `ddlx_create_collation`
 - added `IF NOT EXISTS` in a few more places
+- new options `nodcl`,`noowner`,`nogrants`,`noalter`
 
 Version 0.20
 ------------